+++ src/Bodoconsult.Text/Documents/Blocks/DocumentMetaData.cs
@@ -110,6 +110,16 @@ public class DocumentMetaData : Block, ITypoMetaData
     public string PageNumberPrefix { get; set; } = "Page";
 
     /// <summary>
+    /// Path to the background image or null if no background image should be used. For a A4 portrait page the size of the image should be 1400 px x 1980 px
+    /// </summary>
+    public string BackgroundImagePath { get; set; }
+
+    /// <summary>
+    /// Watermark text
+    /// </summary>
+    public string WatermarkText { get; set; }
+
+    /// <summary>
     /// Prefix for equations
     /// </summary>
     public string EquationPrefix { get; set; } = "Equation";