+++ doc/Bodoconsult.I18N/ReleaseNotes.md
@@ -8,3 +8,9 @@ Release notes for Bodoconsult.I18N
 - 	Using Bodoconsult.App 1.0.8
 
 -	Added I18NServer and I18NClient class for multi-threaded environments with different language requirements per thread
+
+# 1.0.9
+
+- 	Using Bodoconsult.App.Abstractions 1.0.9
+
+-   New interface II18NBase as common base interface for I18N and I18NServer class to make ILocalesProviderPackage more flexible